Move upload/download file methods to the TelegramClient
